requirements.txt
mysql-connector==2.2.9
boto==2.49.0
discord.py==1.5.1
